Add LICENSE
What
===
Add LICENSE containing the standard MIT license sourced from
https://opensource.org/licenses/MIT. The license was sourced from the
gemspec, the year was been sourced from the first commit of the
repository and copyright holder has been sourced from the author name
of the gemspec.

Why
===
The license is specified in the gemspec as the MIT license, but there
is no LICENSE file in the repository. It is standard for GitHub
repositories to contain their license in the LICENSE file, and this
helps inform consumers of the gem, especially those using it in
commerical works.